+++ src/main/java/de/markusbordihn/easynpc/client/screen/components/SaveButton.java
@@ -1,0 +1,58 @@
+/**
+ * Copyright 2023 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package de.markusbordihn.easynpc.client.screen.components;
+
+import java.util.function.Supplier;
+
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
+
+import net.minecraft.client.gui.components.Button;
+import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.network.chat.Component;
+
+import de.markusbordihn.easynpc.Constants;
+
+public class SaveButton extends Button {
+
+  protected static final Button.CreateNarration DEFAULT_NARRATION = Supplier::get;
+
+  public SaveButton(int left, int top, OnPress onPress) {
+    this(left, top, 22, 20, Component.literal(""), onPress);
+  }
+
+  public SaveButton(int left, int top, int width, int height, Component component,
+      OnPress onPress) {
+    super(left, top, width, height, component, onPress, DEFAULT_NARRATION);
+  }
+
+  @Override
+  public void renderWidget(PoseStack poseStack, int left, int top, float partialTicks) {
+    super.renderWidget(poseStack, left, top, partialTicks);
+
+    RenderSystem.setShader(GameRenderer::getPositionTexShader);
+    RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+    RenderSystem.setShaderTexture(0, Constants.TEXTURE_CONFIGURATION);
+
+    // Button Icons
+    blit(poseStack, this.getX() + 4, this.getY() + 3, 60, this.active ? 0 : 16, 16, 16);
+  }
+
+}