+++ src/main/java/de/markusbordihn/easynpc/config/CommonConfig.java
@@ -1,5 +1,61 @@
+/**
+ * Copyright 2023 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
 package de.markusbordihn.easynpc.config;
 
-public class CommonConfig {
+import org.apache.commons.lang3.tuple.Pair;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
+import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.config.ModConfig;
+
+import de.markusbordihn.easynpc.Constants;
+
+@EventBusSubscriber(bus = EventBusSubscriber.Bus.MOD)
+public class CommonConfig {
+  protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+
+  private CommonConfig() {}
+
+  public static final ForgeConfigSpec commonSpec;
+  public static final Config COMMON;
+
+  static {
+    com.electronwill.nightconfig.core.Config.setInsertionOrderPreserved(true);
+    final Pair<Config, ForgeConfigSpec> specPair =
+        new ForgeConfigSpec.Builder().configure(Config::new);
+    commonSpec = specPair.getRight();
+    COMMON = specPair.getLeft();
+    log.info("{} Common config ...", Constants.LOG_REGISTER_PREFIX);
+    ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, commonSpec);
+  }
+
+  public static class Config {
+
+    Config(ForgeConfigSpec.Builder builder) {
+      builder.comment(Constants.MOD_NAME);
+
+      builder.push("General");
+      builder.pop();
+    }
+  }
 }